Move the location of the review list
