+++ packages/review/src/components/shorten-list/popular-reviews.tsx
@@ -1,7 +1,8 @@
 import { usePopularReviews } from '../../services'
-import { ReviewsList } from '../reviews-list'
+
 import type { SortingType } from '../sorting-context'
 
+import { ReviewsList } from './reviews-list'
 import type { ShortenReview } from './types'
 
 export function PopularReviews({
