+++ packages/review/src/components/shorten-list/rating-reviews.tsx
@@ -1,7 +1,8 @@
 import { useRatingReviews } from '../../services'
-import { ReviewsList } from '../reviews-list'
+
 import type { SortingType } from '../sorting-context'
 
+import { ReviewsList } from './reviews-list'
 import type { ExtendShortenReview } from './types'
 
 export function RatingReviews({
